+++ config.example.py
@@ -49,6 +49,7 @@ WATCH_TASKS = [
         "notify": True,
         "max_price": None,
         "blocked_title_keywords": None,
+        "limit": None,
         "enabled": True,
     },
     {
@@ -61,6 +62,7 @@ WATCH_TASKS = [
         "notify": True,
         "max_price": None,
         "blocked_title_keywords": None,
+        "limit": None,
         "enabled": True,
     },
     {
@@ -73,6 +75,7 @@ WATCH_TASKS = [
         "notify": False,
         "max_price": None,
         "blocked_title_keywords": None,
+        "limit": 15,
         "enabled": False,
     },
     {
@@ -85,6 +88,7 @@ WATCH_TASKS = [
         "notify": False,
         "max_price": None,
         "blocked_title_keywords": None,
+        "limit": 15,
         "enabled": False,
     },
     {
@@ -97,6 +101,7 @@ WATCH_TASKS = [
         "notify": True,
         "max_price": None,
         "blocked_title_keywords": None,
+        "limit": 15,
         "enabled": False,
     },
 ]